+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Refactoring of duration format functions (#663)
 -   Wording update for deleted models in output (#669)
+-   Refactoring of DrawerHeader extra buttons (#676)
 
 ## [0.29.0] - 2022-07-25